Not emitting duplicated stack trace. This change makes raise or raise-continuable and saving raised condition a bit more expensive than before (and may cause memory explosion). The basic idea of this change is that using continuation frame as a stack trace so that nested stack trace can be detected easily. To make this happen, we save all frames into the heap as if call/cc is called when raise/raise-contiuable is called.
Above sounds kinda horrible however if we implement segmented stacks type
call/cc described the blow paper, then this performance penalty wouldn't be
a problem.
 - Representing Control in the Presence of First-Class Continuations
   URL: http://www.cs.indiana.edu/~dyb/papers/stack.ps
Not sure if this happens in near future but we may review the current
implementation of call/cc if the performance would be an issue.